Add button under the cards to bring up dialog box
Added a button to pop up dialog box to get info
for the new book to the library. Used dialog
element and put the form inside it that has
label and input text. set inset=0 and margin=auto
to place the form in the middle of the page.
Added two buttons submit and cancel at the end
to send the data to myLibrary array list.
[Ticket: 4]