Add remove & read/notRread button to each card
To connect the DOM element to the actual book
objects in the myLibrary array, the for each loop
is used to associate index to each specific card.
then by using eventListener for the buttons
the index of any card clicked is retrieved.

for read/notRread button I used if statement to
toggle between read and notRread if the button
is clicked. I update them both in the object
attribute and the display.

[Ticket: 5]